added safety checks,
Not working with dirs in single dir

+++ sync.py
@@ -28,16 +28,17 @@ Recursively crawl through all folders calling main
 """
 
 
-def crawl_folders(dir, other_dir):
-    if os.path.exists(dir):  # ie folder exists
-        dir_filenames = os.listdir(dir)
+def crawl_folders(dir1, other_dir):
+    if os.path.exists(dir1):  # ie folder exists
+        dir_filenames = os.listdir(dir1)
         dir2_filenames = os.listdir(other_dir)
         for filename_dir2 in dir2_filenames:
             if any(filename_dir2 in filename_dir1 for filename_dir1 in dir_filenames):
                 dir_filenames.append(filename_dir2)
+        dir_filenames[:] = [name for name in dir_filenames if name != ".sync"]
         for filename in dir_filenames:
             # check if folder and if it is recall main function
-            dir_path1 = os.path.join(dir, filename)
+            dir_path1 = os.path.join(dir1, filename)
             dir_path2 = os.path.join(other_dir, filename)
             if os.path.isdir(dir_path1) or os.path.isdir(dir_path2):
                 if not os.path.exists(dir_path1):
@@ -107,6 +108,9 @@ def update_sync(dir, sync):
     dir_filenames = load_files(dir)
     for filename in dir_filenames:
         pathname = os.path.join(dir, filename)
+        # if os.path.isdir(pathname):
+        #         dir_filenames.remove(filename)
+        #         continue
         if filename != ".sync":
             if filename in sync:
                 file_array = sync.get(filename)
@@ -166,12 +170,14 @@ def sync(sync1, sync2, dir1, dir2):
             if repr(latest_file_update_sync1[0]) > repr(sync2[filename][0][0]) and latest_file_update_sync1[
                 1] == "deleted":
                 pathname = os.path.join(dir2, filename)
-                os.remove(pathname)
+                if os.path.exists(pathname):
+                    os.remove(pathname)
                 continue
             # file deleted in sync 2 and most recent
             elif repr(latest_file_update_sync1[0]) < repr(sync2[filename][0][0]) and sync2[filename][0][1] == "deleted":
                 pathname = os.path.join(dir1, filename)
-                os.remove(pathname)
+                if os.path.exists(pathname):
+                    os.remove(pathname)
                 continue
                 # file deleted in sync 1 or 2 and not most recent
             elif latest_file_update_sync1[1] == "deleted" and repr(latest_file_update_sync1[0]) < repr(
@@ -183,7 +189,8 @@ def sync(sync1, sync2, dir1, dir2):
         # file not in other sync
         if filename not in sync2:
             pathname = os.path.join(dir1, filename)
-            shutil.copy2(pathname, dir2)
+            if os.path.exists(pathname):
+                shutil.copy2(pathname, dir2)
             continue
         # file in other sync then copy the latest of thw two to the other side
         # if a file has the same current digest in both directories but different modification dates
@@ -197,7 +204,8 @@ def sync(sync1, sync2, dir1, dir2):
                     if latest_file_update_sync1[1] == sync2_file_arrays[1]:
                         # file in dir2 must be newer
                         pathname = os.path.join(dir2, filename)
-                        shutil.copy2(pathname, dir1)
+                        if os.path.exists(pathname):
+                            shutil.copy2(pathname, dir1)
                         file_moved_boolean = True
                         break
                     if not file_moved_boolean:
@@ -205,7 +213,8 @@ def sync(sync1, sync2, dir1, dir2):
                             if sync2[filename][0][1] == sync1_file_arrays[1]:
                                 # file in dir1 must be newer
                                 pathname = os.path.join(dir1, filename)
-                                shutil.copy2(pathname, dir2)
+                                if os.path.exists(pathname):
+                                    shutil.copy2(pathname, dir2)
                                 break
                     # we get here only if both digests are unique
                     # if a file has different digests in both directories then the files are different, copy over latest (not ideal)
@@ -222,10 +231,12 @@ def sync(sync1, sync2, dir1, dir2):
 def copy_latest_file(dir1, dir2, filename, latest_file_update_sync1, latest_file_update_sync2):
     if repr(latest_file_update_sync1[0]) > repr(latest_file_update_sync2[0]):
         pathname = os.path.join(dir1, filename)
-        shutil.copy2(pathname, dir2)
+        if os.path.exists(pathname):
+            shutil.copy2(pathname, dir2)
     else:
         pathname = os.path.join(dir2, filename)
-        shutil.copy2(pathname, dir1)
+        if os.path.exists(pathname):
+            shutil.copy2(pathname, dir1)
 
 
 """
@@ -238,5 +249,11 @@ if __name__ == "__main__":
     if len(sys.argv) < 2 or len(sys.argv) > 3:
         print("Usage: sync directory1 directory2")
         exit()
+    if not os.path.exists(sys.argv[1]):  # ie folder exists
+        # create folder
+        os.makedirs(sys.argv[1])
+    if not os.path.exists(sys.argv[2]):  # ie folder exists
+        # create folder
+        os.makedirs(sys.argv[2])
 
     main(sys.argv[1], sys.argv[2])
